README.md
This project is done as a part of my Compiler design Lab.
The aim of this project is to construct a Parsing table for any grammar
sequence entered by the user by finding the FIRST and FOLLOW sets of
each non-terminal and then construct the LL1 parsing table for the same.
The user will have a Graphical Interface to enter the values